imm-automation/SERVER/requirements.txt
redis==4.3.4
cryptography==3.3.2
RPi.GPIO==0.7.1
celery==5.2.7
numpy==1.24.1
asyncua==1.0.1
asgiref==3.6.0
Django==4.1.7
django_debug_toolbar==3.8.1
sqlparse==0.4.3
tzdata